refactor (member): refactor member list folder structure
